+++ R/utils.R
@@ -17,6 +17,16 @@ length_at_age <- function(population, age) {
   })
 }
 
+age_at_length <- function(population, length) {
+  with(population, {
+    if(L2 < 0) L2 <- Linf * (1 - exp(-k * (-L2 - t0)))
+    t <- -log(1 - pmin(length / Linf, 1)) / k - t0
+    t2 <- -log(1 - pmin(L2 / Linf, 1)) / k - t0
+    t[t > t2] <- -log(1 - pmin((length[t > t2] - L2) / (Linf2 - L2), 1)) / k2 + t2
+    t
+  })
+}
+
 #' Length At Age
 #'
 #' @inheritParams params
@@ -30,6 +40,25 @@ ypr_length_at_age <- function(population, age) {
   chk_population(population)
   chk_numeric(age)
   chk_vector(age)
+  chk_gte(age)
 
   length_at_age(population, age)
+}
+
+#' Age At Length
+#'
+#' @inheritParams params
+#'
+#' @return A double vector of the lengths.
+#' @export
+#'
+#' @examples
+#' ypr_age_at_length(ypr_population(), seq(0,100, by = 10))
+ypr_age_at_length <- function(population, length) {
+  chk_population(population)
+  chk_numeric(length)
+  chk_vector(length)
+  chk_gte(length)
+
+  age_at_length(population, length)
 }